topaz: audio: Pull FM Tuner device from audio policy
This is the cause of the "duplicate" device mic shown in recorder (or not at all, making the mic seem "broken"). Since no one is willing to add FM Radio support anyway (and I don't think this is the path nowadays AFAIK), pull this device.

This removes the duplicate microphone and puts the proper internal mic as 'preferred' when no device is connected.

Test: Boots, microphone selection in Google Recorder

Signed-off-by: John Vincent Corcega <git@tenseventyseven.xyz>